# graywolf/pkg/webapi/docs/gen/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API for graywolf configuration and control.",
        "title": "Graywolf Management API",
        "contact": {},
        "version": "1.0"
    },
...
    "basePath": "/api",
    "paths": {
...
        "/messages": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "List messages",
                "operationId": "listMessages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Folder filter: inbox|sent|all",
                        "name": "folder",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "PeerCall filter",
                        "name": "peer",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "dm|tactical",
                        "name": "thread_kind",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Exact thread key (peer callsign for DM, tactical label for tactical)",
                        "name": "thread_key",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Only messages at or after this RFC3339 timestamp",
                        "name": "since",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Opaque cursor from a prior response; pages forward",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Restrict to unread rows",
                        "name": "unread_only",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Cap result count (1..500)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.MessageListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/webtypes.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/webtypes.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "Send message",
                "operationId": "sendMessage",
                "parameters": [
                    {
                        "description": "Compose payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SendMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/dto.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/webtypes.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/webtypes.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/webtypes.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/status": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "status"
                ],
                "summary": "System status dashboard",
                "operationId": "getStatus",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/webapi.StatusDTO"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.MessageChange": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "kind": {
                    "type": "string"
                },
                "message": {
                    "$ref": "#/definitions/dto.MessageResponse"
                }
            }
        },
        "dto.MessageListResponse": {
            "type": "object",
            "properties": {
                "changes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.MessageChange"
                    }
                },
                "cursor": {
                    "type": "string"
                }
            }
        },
...
        "dto.MessageResponse": {
            "type": "object",
            "properties": {
                "acked_at": {
                    "type": "string"
                },
                "attempts": {
                    "type": "integer"
                },
                "channel": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "direction": {
                    "description": "\"in\" | \"out\"",
                    "type": "string"
                },
                "failure_reason": {
                    "type": "string"
                },
                "from_call": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_ack": {
                    "type": "boolean"
                },
                "is_bulletin": {
                    "type": "boolean"
                },
                "msg_id": {
                    "type": "string"
                },
                "next_retry_at": {
                    "type": "string"
                },
                "our_call": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "peer_call": {
                    "type": "string"
                },
                "received_at": {
                    "type": "string"
                },
                "received_by_call": {
                    "type": "string"
                },
                "sent_at": {
                    "type": "string"
                },
                "source": {
                    "description": "\"rf\" | \"is\"",
                    "type": "string"
                },
                "status": {
                    "description": "derived — see DeriveMessageStatus",
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "thread_key": {
                    "type": "string"
                },
                "thread_kind": {
                    "type": "string"
                },
                "to_call": {
                    "type": "string"
                },
                "unread": {
                    "type": "boolean"
                },
                "via": {
                    "type": "string"
                }
            }
        },
...
        "dto.SendMessageRequest": {
            "type": "object",
            "properties": {
                "channel": {
                    "description": "Channel overrides the configured TX channel. Nil = use default.",
                    "type": "integer"
                },
                "client_id": {
                    "description": "ClientID is an opaque client-side correlation token. Echoed back\nunchanged in the response so the optimistic UI can reconcile its\nlocal row with the persisted ID.",
                    "type": "string"
                },
                "path": {
                    "description": "Path overrides the default RF path from preferences. Empty =\nuse MessagePreferences.DefaultPath.",
                    "type": "string"
                },
                "prefer_is": {
                    "description": "PreferIS, when true, routes the outbound via APRS-IS regardless\nof the current fallback policy.",
                    "type": "boolean"
                },
                "text": {
                    "description": "Text is the message body (\u003c= 67 APRS chars after validation).",
                    "type": "string"
                },
                "to": {
                    "description": "To is the addressee: a station callsign for a DM or a tactical\nlabel for a group broadcast. Uppercase-normalized server-side.",
                    "type": "string"
                }
            }
        },
...
        "webapi.StatusChannel": {
            "type": "object",
            "properties": {
                "audio_peak": {
                    "type": "number"
                },
                "bit_rate": {
                    "type": "integer"
                },
                "dcd_state": {
                    "type": "boolean"
                },
                "device_clipping": {
                    "type": "boolean"
                },
                "device_peak_dbfs": {
                    "type": "number"
                },
                "device_rms_dbfs": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "input_device_id": {
                    "type": "integer"
                },
                "modem_type": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "rx_frames": {
                    "type": "integer"
                },
                "tx_frames": {
                    "type": "integer"
                }
            }
        },
        "webapi.StatusDTO": {
            "type": "object",
            "properties": {
                "channels": {
                    "description": "Channels is the per-channel live status snapshot (frame counters, audio levels).",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/webapi.StatusChannel"
                    }
                },
                "igate": {
                    "description": "Igate is the current iGate session state; omitted when no iGate is configured.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/webapi.StatusIgateDTO"
                        }
                    ]
                },
                "uptime_seconds": {
                    "description": "UptimeSeconds is the server process uptime in whole seconds.",
                    "type": "integer"
                }
            }
        },
        "webapi.StatusIgateDTO": {
            "type": "object",
            "properties": {
                "callsign": {
                    "description": "Callsign is the login callsign-SSID presented to APRS-IS.",
                    "type": "string"
                },
                "connected": {
                    "description": "Connected is true while an APRS-IS session is established.",
                    "type": "boolean"
                },
                "is_to_rf_gated": {
                    "description": "Downlinked is the cumulative count of IS packets transmitted on RF.",
                    "type": "integer"
                },
                "last_connected": {
                    "description": "LastConnected is the UTC RFC3339 timestamp of the most recent successful IS login; omitted if never connected.",
                    "type": "string"
                },
                "packets_filtered": {
                    "description": "Filtered is the cumulative count of packets dropped by the filter engine.",
                    "type": "integer"
                },
                "rf_to_is_dropped": {
                    "description": "DroppedOffline is the cumulative count of RF packets dropped because the IS session was offline.",
                    "type": "integer"
                },
                "rf_to_is_gated": {
                    "description": "Gated is the cumulative count of RF packets forwarded to APRS-IS.",
                    "type": "integer"
                },
                "server": {
                    "description": "Server is the APRS-IS host:port currently in use (or last attempted).",
                    "type": "string"
                },
                "simulation_mode": {
                    "description": "SimulationMode is true when RF-\u003eIS uploads are suppressed for testing.",
                    "type": "boolean"
                }
            }
        },
...
        "webtypes.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "CookieAuth": {
            "description": "Session cookie issued by POST /api/auth/login.\nSwagger 2.0 lacks a native `in: cookie` apiKey\nlocation, so the spec models the same credential\nas a `Cookie:` request header. Browsers send this\nautomatically once the session cookie is set; the\nsession cookie is named `graywolf_session`.",
            "type": "apiKey",
            "name": "Cookie",
            "in": "header"
        }
    }
}